First 'hacky' implementation to address issue #129
We check if the 'current' window is iframed. If not we keep the usual logic.
Otherwise, we pop a new window with the right url to https://login.microsoftonline.com/etc..., when the url is changed (after clicking SignUp)
we close the popup and use this new url for the parent iframe.